requirements.txt
# Streamlit para la interfaz
streamlit

# Requests para API
requests

# Manage the enviroments
python-dotenv